cmd: split cmd into separate binaries
